pyproject.toml
[project]
name = "eqemu-mcp-server"
version = "0.4.0"
description = "MCP server for EQEmu — C++ source search, quest API, database, NPC/item/spawn/loot management, documentation"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "mcp[cli]>=1.0.0",
    "mysql-connector-python>=8.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
